moved the actual office retrieval into a seperate object to group the database code.
